requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==1.11.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==4.2.16
django-emoji==2.2.2
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==1.0.0
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
lxml==5.3.0
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.24.4
openai==1.47.1
outcome==1.3.0.post0
pandas==2.0.3
pillow==10.4.0
prettytable==2.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2024.1
redis==5.1.1
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spotify==0.10.2
spotipy==2.24.0
SQLAlchemy==2.0.36
sqlparse==0.5.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
twilio==9.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
visitor==0.1.3
wcwidth==0.2.5
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
yarl==1.11.1
zipp==3.20.2
python-dotenv==1.0.1
